corrected a few g++ warnings with -Wall
git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@3523 2422e80f-9519-0410-a819-8e21e7c86b6d

+++ src/PSFCatalog.cpp
@@ -609,7 +609,7 @@ void PSFCatalog::ReadFits(std::string file)
     std::valarray<double> coeffs;
     table.column(coeffs_col).read(coeffs, row);
     double* ptri = (double* ) psf[i].cptr(); 
-    for (size_t j=0; j<ncoeff; ++j) {
+    for (int j=0; j<ncoeff; ++j) {
       ptri[i] = coeffs[i];
     }
   }
